+++ src/main/python/apache/aurora/executor/common/health_checker.py
@@ -208,6 +208,9 @@ class HealthChecker(StatusChecker):
 
 class HealthCheckerProvider(StatusCheckerProvider):
 
+  def __init__(self, setuid_health_checks=True):
+    self.setuid_health_checks = setuid_health_checks
+
   @staticmethod
   def interpolate_cmd(task, cmd):
     """
@@ -244,14 +247,17 @@ class HealthCheckerProvider(StatusCheckerProvider):
         task=assigned_task,
         cmd=shell_command
       )
-
-      pw_entry = pwd.getpwnam(assigned_task.task.job.role)
-      def demote_to_user():
-        os.setgid(pw_entry.pw_gid)
-        os.setuid(pw_entry.pw_uid)
+      # If we do not want user which is job's role to execute the health shell check.
+      if self.setuid_health_checks:
+        pw_entry = pwd.getpwnam(assigned_task.task.job.role)
+        def demote_to_job_role_user():
+          os.setgid(pw_entry.pw_gid)
+          os.setuid(pw_entry.pw_uid)
+      else:
+        demote_to_job_role_user = None
 
       shell_signaler = ShellHealthCheck(cmd=interpolated_command,
-        preexec_fn=demote_to_user,
+        preexec_fn=demote_to_job_role_user,
         timeout_secs=timeout_secs)
       a_health_checker = lambda: shell_signaler()
     else: