+++ src/test/python/apache/aurora/executor/common/test_health_checker.py
@@ -282,6 +282,42 @@ class TestHealthCheckerProvider(unittest.TestCase):
     assert hct_max_fail == max_consecutive_failures
     mock_getpwnam.assert_called_once_with(task_config.job.role)
 
+  @mock.patch('pwd.getpwnam')
+  def test_from_assigned_task_shell_no_demotion(self, mock_getpwnam):
+    interval_secs = 17
+    initial_interval_secs = 3
+    max_consecutive_failures = 2
+    timeout_secs = 5
+    shell_config = ShellHealthChecker(shell_command='failed command')
+    task_config = TaskConfig(
+        job=JobKey(role='role', environment='env', name='name'),
+        executorConfig=ExecutorConfig(
+            name='thermos-generic',
+            data=MESOS_JOB(
+                task=HELLO_WORLD,
+                health_check_config=HealthCheckConfig(
+                    health_checker=HealthCheckerConfig(shell=shell_config),
+                    interval_secs=interval_secs,
+                    initial_interval_secs=initial_interval_secs,
+                    max_consecutive_failures=max_consecutive_failures,
+                    timeout_secs=timeout_secs,
+                )
+            ).json_dumps()
+        )
+    )
+    assigned_task = AssignedTask(task=task_config, instanceId=1, assignedPorts={'foo': 9001})
+    execconfig_data = json.loads(assigned_task.task.executorConfig.data)
+    assert execconfig_data[
+             'health_check_config']['health_checker']['shell']['shell_command'] == 'failed command'
+    health_checker = HealthCheckerProvider(setuid_health_checks=False).from_assigned_task(
+      assigned_task, None)
+    assert health_checker.threaded_health_checker.interval == interval_secs
+    assert health_checker.threaded_health_checker.initial_interval == initial_interval_secs
+    hct_max_fail = health_checker.threaded_health_checker.max_consecutive_failures
+    assert hct_max_fail == max_consecutive_failures
+    # Should not be trying to access role's user info.
+    assert not mock_getpwnam.called
+
   def test_interpolate_cmd(self):
     """Making sure thermos.ports[foo] gets correctly substituted with assignedPorts info."""
     interval_secs = 17